Fixed new isssue

+++ controller/Patients.js
@@ -15,154 +15,26 @@ function deleteFile(path) {
   }
 }
 
-
 const PatientsController = {
   async addPatient(req, res, next) {
-    
-    upload.fields([{ name: "image" }, { name: "pdffile" }])(req, res, async function (err) {
-      if (err) {
-        return next(err);
-      }
-  
-      const validateData = Joi.object({
-        patient_name: Joi.string().required(),
-        admit_date: Joi.date().required(),
-        patient_cnic: Joi.number().required(),
-        patient_case_type: Joi.string().required(),
-        discharg_date: Joi.string(),
-
-        user: Joi.string().regex(myregexp).required(),
-      });
-  
-      const { error } = validateData.validate(req.body);
-      if (error) {
-        // Delete files if they were uploaded before validation failed
-        if (req.files && req.files.image) {
-          deleteFile(req.files.image[0].path);
-        }
-        if (req.files && req.files.pdffile) {
-          deleteFile(req.files.pdffile[0].path);
-        }
-        return next(error);
-      }
-  
-      const {
-        patient_name,
-        admit_date,
-        patient_cnic,
-        patient_case_type,
-        discharg_date,
-        user,
-      } = req.body;
-  
-      let imagePath, pdfFilePath;
-  
-      // Handle image upload
-      if (req.files && req.files.image) {
-        const imageFile = req.files.image[0];
-        const imageFolder = "images";
-        imagePath = `${imageFolder}/image-${Date.now()}` + `-patientimage.png`;
-  
-        try {
-          // Ensure the folder exists
-          if (!fs.existsSync(`storage/${imageFolder}`)) {
-            fs.mkdirSync(`storage/${imageFolder}`);
-          }
-  
-          fs.writeFileSync(`storage/${imagePath}`, imageFile.buffer);
-        } catch (error) {
-          // Delete the file if an error occurs
-          deleteFile(imageFile.path);
-          return next(error);
-        }
-      }
-  
-      // Handle pdffile upload
-      if (req.files && req.files.pdffile) {
-        const pdfFile = req.files.pdffile[0];
-        const pdfFolder = "pdfs";
-        const pdfFileName = `pdf-${Date.now()}-file.pdf`;
-        pdfFilePath = path.join("storage", pdfFolder, pdfFileName);
-  
-        try {
-          // Ensure the folder exists
-          if (!fs.existsSync(`storage/${pdfFolder}`)) {
-            fs.mkdirSync(`storage/${pdfFolder}`);
-          }
-  
-          fs.writeFileSync(pdfFilePath, pdfFile.buffer);
-        } catch (error) {
-          // Delete the file if an error occurs
-          deleteFile(pdfFile.path);
-          return next(error);
-        }
-      }
-  
-      let newPatientData;
-      try {
-        newPatientData = new Patients({
-          patient_name,
-          admit_date,
-          patient_cnic,
-          patient_case_type,
-          discharg_date:discharg_date?discharg_date:null,
-          user,
-          image: imagePath ? `${BACKEND_URL_PATH}storage/${imagePath}` : null,
-          pdffile: pdfFilePath ? `${BACKEND_URL_PATH}${pdfFilePath}` : null,
-        });
-        await newPatientData.save();
-      } catch (err) {
-        // Delete the files if an error occurs
-        deleteFile(`storage/${imagePath}`);
-        deleteFile(pdfFilePath);
-        return next(err);
-      }
-  
-      const newPatient = new PatientDto(newPatientData);
-      return res.status(201).json({ patient: newPatient });
-    });
-    
-  },
-
-  async getPatients(req, res, next) {
-    let newPatientData;
-    try {
-
-      if(req.user.role.toLowerCase() !=="admin" ){
-        newPatientData = await Patients.find({patient_case_type:req.user.role.toLowerCase()});
-      } else {
-        newPatientData = await Patients.find();
-      }
-        return res.status(200).json({ patients: newPatientData?.length ?newPatientData:[]  });
-     
-    } catch (err) {
-      return next(err);
-    }
-  },
-  async updatePatient(req, res, next) {
-    try {
-  
-      const patientId = req.params.patientId;
-      const existingPatient = await Patients.findById(patientId);
-
-      if (!existingPatient) {
-        return res.status(404).json({ error: 'Patient not found' });
-      }
-  
-      upload.fields([{ name: 'image' }, { name: 'pdffile' }])(req, res, async function (err) {
+    upload.fields([{ name: "image" }, { name: "pdffile" }])(
+      req,
+      res,
+      async function (err) {
         if (err) {
           return next(err);
         }
-  
+
         const validateData = Joi.object({
           patient_name: Joi.string().required(),
           admit_date: Joi.date().required(),
           patient_cnic: Joi.number().required(),
           patient_case_type: Joi.string().required(),
           discharg_date: Joi.string(),
+
           user: Joi.string().regex(myregexp).required(),
         });
-  
+
         const { error } = validateData.validate(req.body);
         if (error) {
           // Delete files if they were uploaded before validation failed
@@ -174,7 +46,7 @@ const PatientsController = {
           }
           return next(error);
         }
-  
+
         const {
           patient_name,
           admit_date,
@@ -183,56 +55,59 @@ const PatientsController = {
           discharg_date,
           user,
         } = req.body;
-  
-        let imagePath = existingPatient.image;
-        let pdfFilePath = existingPatient.pdffile;
-  
+
+        let imagePath, pdfFilePath;
+
         // Handle image upload
         if (req.files && req.files.image) {
           const imageFile = req.files.image[0];
-          const imageFolder = 'images';
-          const oldImageFilename = imagePath ? imagePath.replace(`${BACKEND_URL_PATH}storage/${imageFolder}/`, '') : null;
-          imagePath = `${imageFolder}/image-${Date.now()}` + `-patientimage.png`;
-  
+          if (
+            imageFile.mimetype !== "image/jpeg" &&
+            imageFile.mimetype !== "image/png"
+          ) {
+            return next(new Error("Invalid image format"));
+          }
+
+          const imageFolder = "images";
+          imagePath =
+            `${imageFolder}/image-${Date.now()}` + `-patient-image.png`;
+
           try {
             // Ensure the folder exists
             if (!fs.existsSync(`storage/${imageFolder}`)) {
               fs.mkdirSync(`storage/${imageFolder}`);
             }
-  
-            // Delete the old image file if it exists
-            if (oldImageFilename) {
-              deleteFile(path.join(__dirname, '..', 'storage', imageFolder, oldImageFilename));
-            }
-  
+
             fs.writeFileSync(`storage/${imagePath}`, imageFile.buffer);
           } catch (error) {
-            // Delete the file if an error occurs
             deleteFile(imageFile.path);
             return next(error);
           }
         }
-  
+
         // Handle pdffile upload
         if (req.files && req.files.pdffile) {
           const pdfFile = req.files.pdffile[0];
-          const pdfFolder = 'pdfs';
-        
+          const isPdf = req.files.pdffile[0].mimetype === "application/pdf";
+          if (!isPdf) {
+            return next(new Error("Invalid pdf format"));
+          }
+
+          const pdfFolder = "pdfs";
+          const pdfFileName = `pdf-${Date.now()}.pdf`;
+
+          pdfFilePath = path
+            .join("storage", pdfFolder, pdfFileName)
+            .replace(/\\/g, "/");
+
           try {
             // Ensure the folder exists
-            if (!fs.existsSync(`storage/${pdfFolder}`)) {
-              fs.mkdirSync(`storage/${pdfFolder}`);
+            if (!fs.existsSync(path.join("storage", pdfFolder))) {
+              fs.mkdirSync(path.join("storage", pdfFolder), {
+                recursive: true,
+              });
             }
-        
-            // Delete the old pdf file if it exists
-            if (pdfFilePath) {
-              deleteFile(path.join(__dirname, '..', 'storage', pdfFolder, pdfFilePath));
-            }
-        
-            // Generate a new unique filename
-            const pdfFileName = `pdf-${Date.now()}-file.pdf`;
-            pdfFilePath = path.join('storage', pdfFolder, pdfFileName);
-        
+
             fs.writeFileSync(pdfFilePath, pdfFile.buffer);
           } catch (error) {
             // Delete the file if an error occurs
@@ -240,28 +115,235 @@ const PatientsController = {
             return next(error);
           }
         }
-  
-        // Update patient record with the new data
-        // Update patient record with the new data
-      existingPatient.patient_name = patient_name;
-      existingPatient.admit_date = admit_date;
-      existingPatient.patient_cnic = patient_cnic;
-      existingPatient.patient_case_type = patient_case_type;
-      existingPatient.discharg_date = discharg_date ? discharg_date : existingPatient.discharg_date;
-      existingPatient.user = user;
-      existingPatient.image = imagePath ? `${BACKEND_URL_PATH}storage/${imagePath}` : null;
-      existingPatient.pdffile = pdfFilePath ? `${BACKEND_URL_PATH}${pdfFilePath}` : null;
 
-      await existingPatient.save();
-  
-        const updatedPatient = new PatientDto(existingPatient);
-        return res.status(200).json({ patient: updatedPatient });
-      });
+        let newPatientData;
+        try {
+          newPatientData = new Patients({
+            patient_name,
+            admit_date,
+            patient_cnic,
+            patient_case_type,
+            discharg_date: discharg_date ? discharg_date : null,
+            user,
+            image: imagePath ? `${BACKEND_URL_PATH}storage/${imagePath}` : null,
+            pdffile: pdfFilePath ? `${BACKEND_URL_PATH}${pdfFilePath}` : null,
+          });
+          await newPatientData.save();
+        } catch (err) {
+          // Delete the files if an error occurs
+          deleteFile(`storage/${imagePath}`);
+          deleteFile(pdfFilePath);
+          return next(err);
+        }
+
+        const newPatient = new PatientDto(newPatientData);
+        return res.status(201).json({ patient: newPatient });
+      }
+    );
+  },
+
+  async getPatients(req, res, next) {
+    let newPatientData;
+    try {
+      if (req.user.role.toLowerCase() !== "admin") {
+        newPatientData = await Patients.find({
+          patient_case_type: req.user.role.toLowerCase(),
+        });
+      } else {
+        newPatientData = await Patients.find();
+      }
+      return res
+        .status(200)
+        .json({ patients: newPatientData?.length ? newPatientData : [] });
+    } catch (err) {
+      return next(err);
+    }
+  },
+  async updatePatient(req, res, next) {
+    try {
+      const patientId = req.params.patientId;
+      const existingPatient = await Patients.findById(patientId);
+
+      if (!existingPatient) {
+        return res.status(404).json({ error: "Patient not found" });
+      }
+      upload.fields([{ name: "image" }, { name: "pdffile" }])(
+        req,
+        res,
+        async function (err) {
+          if (err) {
+            return next(err);
+          }
+
+          const validateData = Joi.object({
+            patient_name: Joi.string().required(),
+            admit_date: Joi.date().required(),
+            patient_cnic: Joi.number().required(),
+            patient_case_type: Joi.string().required(),
+            discharg_date: Joi.string(),
+            user: Joi.string().regex(myregexp).required(),
+          });
+
+          const { error } = validateData.validate(req.body);
+          if (error) {
+            // Delete files if they were uploaded before validation failed
+            if (req.files && req.files.image) {
+              deleteFile(req.files.image[0].path);
+            }
+            if (req.files && req.files.pdffile) {
+              deleteFile(req.files.pdffile[0].path);
+            }
+            return next(error);
+          }
+
+          const {
+            patient_name,
+            admit_date,
+            patient_cnic,
+            patient_case_type,
+            discharg_date,
+            user,
+          } = req.body;
+
+          let imagePath = existingPatient.image;
+          let pdfFilePath = existingPatient.pdffile;
+
+          // Handle image upload
+          if (req.files && req.files.image) {
+            const imageFile = req.files.image[0];
+            if (
+              imageFile.mimetype !== "image/jpeg" &&
+              imageFile.mimetype !== "image/png"
+            ) {
+              // Throw an error or send a response informing the user about invalid image format
+              return next(new Error("Invalid image format"));
+            }
+            const imageFolder = "images";
+            const oldImageFilename = imagePath
+              ? imagePath.replace(
+                  `${BACKEND_URL_PATH}storage/${imageFolder}/`,
+                  ""
+                )
+              : null;
+            imagePath =
+              `${imageFolder}/image-${Date.now()}` + `-patient-image.png`;
+
+            try {
+              // Ensure the folder exists
+              const storageFolderPath = path.join(
+                __dirname,
+                "..",
+                "storage",
+                imageFolder
+              );
+              
+              if (!fs.existsSync(storageFolderPath)) {
+                fs.mkdirSync(storageFolderPath, { recursive: true });
+              }
+
+              // Delete the old image file if it exists
+           
+              if (oldImageFilename) {
+                const localImagePath = path
+                  .join("storage", imageFolder,
+                  oldImageFilename)
+                  .replace(/\\/g, "/");
+
+                if (fs.existsSync(localImagePath)) {
+                  deleteFile(localImagePath);
+                } else {
+                  console.log(`File does not exist: ${localImagePath}`);
+                }
+              }
+
+              fs.writeFileSync(`storage/${imagePath}`, imageFile.buffer);
+            } catch (error) {
+              // Delete the file if an error occurs
+              deleteFile(imageFile.path);
+              return next(error);
+            }
+          }
+
+          // Handle pdffile upload
+          if (req.files && req.files.pdffile) {
+            const pdfFile = req.files.pdffile[0];
+            const isPdf = req.files.pdffile[0].mimetype === "application/pdf";
+            if (!isPdf) {
+              return next(new Error("Invalid pdf format"));
+            }
+
+            const pdfFolder = "pdfs";
+            const pdfFileRelativePath = path
+              .join("storage", pdfFolder)
+              .replace(/\\/g, "/");
+
+            try {
+              if (!fs.existsSync(pdfFileRelativePath)) {
+                fs.mkdirSync(pdfFileRelativePath);
+              }
+
+              // Delete the old pdf file if it exists
+              if (pdfFilePath) {
+                const urlParts = pdfFilePath.split("/");
+                const fileName = urlParts[urlParts.length - 1];
+
+                const localFilePath = path
+                  .join("storage/", "pdfs", fileName)
+                  .replace(/\\/g, "/");
+
+                if (fs.existsSync(localFilePath)) {
+                  deleteFile(localFilePath);
+                } else {
+                  console.log(`File does not exist: ${localFilePath}`);
+                }
+              }
+
+              // Generate a new unique filename
+              const pdfFileName = `pdf-${Date.now()}.pdf`;
+              pdfFilePath = path
+                .join("storage", pdfFolder, pdfFileName)
+                .replace(/\\/g, "/");
+
+              fs.writeFileSync(pdfFilePath, pdfFile.buffer);
+            } catch (error) {
+              // Delete the file if an error occurs
+              deleteFile(pdfFile.path);
+              return next(error);
+            }
+          }
+
+          existingPatient.patient_name = patient_name;
+          existingPatient.admit_date = admit_date;
+          existingPatient.patient_cnic = patient_cnic;
+          existingPatient.patient_case_type = patient_case_type;
+          existingPatient.discharg_date = discharg_date
+            ? discharg_date
+            : existingPatient.discharg_date;
+          existingPatient.user = user;
+
+          // Handle image upload
+          if (req.files && req.files.image) {
+            existingPatient.image = imagePath
+              ? `${BACKEND_URL_PATH}storage/${imagePath}`
+              : existingPatient.image;
+          }
+          if (req.files && req.files.pdffile) {
+            existingPatient.pdffile = pdfFilePath
+              ? `${BACKEND_URL_PATH}${pdfFilePath}`
+              : existingPatient.pdffile;
+          }
+
+          await existingPatient.save();
+
+          const updatedPatient = new PatientDto(existingPatient);
+          return res.status(200).json({ patient: updatedPatient });
+        }
+      );
     } catch (error) {
       return next(error);
     }
-  }
-  
+  },
+
   // async postById(req, res, next) {
   //   var validateData = Joi.object({
   //     id: Joi.string().regex(myregexp).required(),
